+++ components/section.tsx
@@ -1,0 +1,13 @@
+type Props = {
+  title: string;
+  children: React.ReactNode;
+};
+
+export default function Section({ title, children }: Props) {
+  return (
+    <div>
+      <p className="font-semibold">{title}</p>
+      <ul>{children}</ul>
+    </div>
+  );
+}